feat(Sep_Generic_Wp) add cost type for space
the type space_cost captures the maximum space used during a computation,
by a pair of numbers denoting the maximum space used (m) in the computation,
and the delta in space consuption (c) from start to end of the computation.
We represent this as a pair of natural numbers (m, m-c) ensuring that in that representation c is at most m.